updated backend project structure
