+++ backend/api/scripts/__init__.py
@@ -1,0 +1,2 @@
+from .db_crud import fetch_all_projects, pop_project
+from .github_api_wrapper import GitHubAPIWrapper
