+++ backend/api/utils/__init__.py
@@ -1,3 +1,1 @@
-from .db_utils import fetch_all_projects
-from .github_wrapper_util import GitHubAPIWrapper 
 from .models import Projects, Users, SavedProjects
